+++ services/gateway/src/types/connection.d.ts
@@ -3,7 +3,7 @@ import * as enums from '../enums';
 
 export interface IRabbitMessage {
   user: IUsersTokens | undefined;
-  target: enums.EUserTargets | enums.EMessageType;
+  target: enums.EUserTargets | enums.EMessageTypes;
   payload: unknown;
 }