+++ services/users/jest.config.e2e.ts
@@ -1,0 +1,9 @@
+import { Config } from 'jest';
+import defaultConfig from './jest.config.default';
+
+const config: Config = {
+  ...defaultConfig,
+  roots: ['./tests/e2e'],
+};
+
+export default config;