+++ services/users/src/types/state.d.ts
@@ -1,0 +1,11 @@
+import Broker from '../broker';
+
+export interface IState {
+  Broker: Broker;
+}
+
+export interface IConfigInterface {
+  amqpURI: string;
+  token: string;
+  refToken: string;
+}